requirements.txt
Flask>=2.0
python-telegram-bot>=20.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=0.21
bcrypt>=4.0
gunicorn
requests
